% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_build_map_url.R
\name{build_map_url}
\alias{build_map_url}
\title{Build a complete map fisbroker url}
\usage{
build_map_url(map_resource)
}
\arguments{
\item{map_resource}{name of a single map_resource ("Kartenservice)
from the WFS service of the FisBroker Berlin}
}
\value{
character
}
\description{
Creates full url for a single map_resource of the FisBroker Berlin that can be
viewed in a Browser.
}
