% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_extract_fisbroker.R
\name{build_wfs_url}
\alias{build_wfs_url}
\title{Build a complete wfs fisbroker url}
\usage{
build_wfs_url(wfs_resource)
}
\arguments{
\item{wfs_resource}{name of a single wfs_resource from the WFS service of the FisBroker Berlin}
}
\value{
character
}
\description{
creates full url for a single wfs_resource WFS service of the FisBroker Berlin
to be passed on to extract_fisbroker()
}
