% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_extract_fisbroker.R
\name{extract_fisbroker}
\alias{extract_fisbroker}
\title{Get fisbroker data}
\usage{
extract_fisbroker(wfs_resource, crs = "EPSG:25833", ...)
}
\arguments{
\item{wfs_resource}{name of a single wfs_resource from the WFS service of the FisBroker Berlin}

\item{crs}{coordinate reference system or spatial reference
system (srs) - defaults to "EPSG:25833"}

\item{...}{arbitrary arguments for parameterized queries.}
}
\value{
sf object
}
\description{
Downloads raw geodata from WFS services of the fisbroker Berlin
}
