% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_transform_fisbroker.R
\name{get_coordinates}
\alias{get_coordinates}
\title{Extract coordinates from a geometry}
\usage{
get_coordinates(sf_obj)
}
\arguments{
\item{sf_obj}{spatial point sf object}
}
\value{
sf object with additional columns "x" and "y"
}
\description{
Adds coordinates (lon/lat) to a sf object.
}
