% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_berlin_lor.R
\name{transform.berlin_lor}
\alias{transform.berlin_lor}
\title{Transform for berlin_lor}
\usage{
\method{transform}{berlin_lor}(x, ...)
}
\arguments{
\item{x}{fisbroker single_source}

\item{...}{other parameters}
}
\description{
Transform for berlin_lor
}
