% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_transform_fisbroker.R
\name{transform_fisbroker}
\alias{transform_fisbroker}
\title{Transforms downloaded fisbroker data}
\usage{
transform_fisbroker(sf_obj)
}
\arguments{
\item{sf_obj}{spatial sf object}
}
\value{
tibble with geometry column as character (geojson).
Geometry has been reprojected to WGS84.
For spatial points additional columns lon/lat (double) will be added.
}
\description{
Transforms downloaded fisbroker data and makes them ready for sending
to database. The transformation includes reprojection to WGS84 (=CRS 4326)
and adding lon/lat as additional columns for spatial points.
Also, the geometry column will be transformed to a geojson character.
}
